omdlib: Move user/group logic to dedicated module
